# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My REST API",
        "description": "This is the OpenAPI documentation for my Laravel REST API project.",
        "contact": {
            "name": "Mantas",
            "email": "mantas@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/car-models": {
            "get": {
                "tags": [
                    "Car Models"
                ],
                "summary": "Get list of all car models",
                "operationId": "e3b490883632ae064d1de5ddac65cc7e",
                "responses": {
                    "200": {
                        "description": "List of car models",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CarModel"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Car Models"
                ],
                "summary": "Create a new car model",
                "operationId": "3484efe17cc38d2cbf85e61aaaee781a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarModelCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Car model created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CarModel"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/brands/{brandId}/car-models": {
            "get": {
                "tags": [
                    "Car Models"
                ],
                "summary": "Get car models for a specific brand",
                "operationId": "63632475888a5d706e372e61f4e25525",
                "parameters": [
                    {
                        "name": "brandId",
                        "in": "path",
                        "description": "ID of the brand",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of car models belonging to the brand",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CarModel"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Brand not found"
                    }
                }
            }
        },
        "/api/car-models/{id}": {
            "get": {
                "tags": [
                    "Car Models"
                ],
                "summary": "Get details of a specific car model",
                "operationId": "64a838d8b35a648d96ece48c74a9d740",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Car model ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Car model details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CarModel"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Car model not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Car Models"
                ],
                "summary": "Update an existing car model",
                "operationId": "2efbf1d7d195e91cd151e6395fab76b3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Car model ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarModelUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Car model updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CarModel"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Car model not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Car Models"
                ],
                "summary": "Delete a car model by ID",
                "operationId": "891983dc0d18b7d798f7ca2459643ab8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Car model ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Car model deleted successfully"
                    },
                    "404": {
                        "description": "Car model not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CarModel": {
                "title": "CarModel",
                "description": "Represents a car model",
                "required": [
                    "id",
                    "brand_id",
                    "name",
                    "year"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "brand_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "name": {
                        "type": "string",
                        "example": "Corolla"
                    },
                    "year": {
                        "type": "integer",
                        "example": 2018
                    },
                    "description": {
                        "type": "string",
                        "example": "Reliable and fuel-efficient sedan"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-14T12:30:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-14T12:45:00Z"
                    }
                },
                "type": "object"
            },
            "CarModelCreateRequest": {
                "title": "CarModelCreateRequest",
                "description": "Request body for creating a CarModel",
                "required": [
                    "brand_id",
                    "name",
                    "year"
                ],
                "properties": {
                    "brand_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "name": {
                        "type": "string",
                        "example": "Corolla"
                    },
                    "year": {
                        "type": "integer",
                        "example": 2018
                    },
                    "description": {
                        "type": "string",
                        "example": "Reliable and fuel-efficient sedan"
                    }
                },
                "type": "object"
            },
            "CarModelUpdateRequest": {
                "title": "CarModelUpdateRequest",
                "description": "Request body for updating a CarModel",
                "required": [
                    "brand_id",
                    "name",
                    "year"
                ],
                "properties": {
                    "brand_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "name": {
                        "type": "string",
                        "example": "Corolla"
                    },
                    "year": {
                        "type": "integer",
                        "example": 2018
                    },
                    "description": {
                        "type": "string",
                        "example": "Reliable and fuel-efficient sedan"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}